OOP Exam Preparation

+++ Preparation/Glacial Expedition/src/main/java/glacialExpedition/models/explorers/AnimalExplorer.java
@@ -1,7 +1,7 @@
 package glacialExpedition.models.explorers;
 
 public class AnimalExplorer extends BaseExplorer{
-    private static final int INITIAL_UNITS_OF_ENERGY=100;
+    private static final double INITIAL_UNITS_OF_ENERGY=100;
     public AnimalExplorer(String name) {
         super(name, INITIAL_UNITS_OF_ENERGY);
     }
