+++ Preparation/Glacial Expedition/src/main/java/glacialExpedition/models/explorers/GlacierExplorer.java
@@ -1,7 +1,7 @@
 package glacialExpedition.models.explorers;
 
 public class GlacierExplorer extends BaseExplorer{
-    private static final int INITIAL_UNITS_OF_ENERGY=40;
+    private static final double INITIAL_UNITS_OF_ENERGY=40;
     public GlacierExplorer(String name) {
         super(name, INITIAL_UNITS_OF_ENERGY);
     }
